Added special handling for a "Default Game Settings" mod
If a mod called "Default Game Settings" is found to be selected for installation to the game, options from the Default Game Settings tab will be ignored when saving. This allows custom game settings to be stored in the Mods Library, rather than relying on the GUI (particularly useful when applying changes to a vanilla DOL).

Using the "Update Default Game Settings Only" option in that tab can override this behavior and force game settings from the GUI to be used instead. Or of course you can just disable/remove the mod from the library.